services/api_gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pyjwt>=2.12.1",
]
